refactor(#3608): composition 헬퍼·테스트를 chunk_compose sibling 모듈로 이동 — hotfile/giant-file ratchet 준수 (동작 불변)
