+++ src/services/discord/turn_bridge/mod.rs
@@ -1,4 +1,5 @@
 mod cancel_finalize_policy;
+mod chunk_compose;
 mod completion_guard;
 mod context_window;
 mod headless_delivery;
@@ -424,193 +425,6 @@ fn spawn_stream_message_receiver_adapter(
         }
     });
     StreamMessageReceiverAdapter { rx: async_rx, stop }
-}
-
-/// #3608: true when the accumulated `full_response` currently ends *inside* an
-/// open ``` code fence. Mirrors the fence toggle used by `format_for_discord`
-/// (`trim_start().starts_with("```")`), so blank-line runs the model placed
-/// inside a fence are treated as intentional and left untouched.
-fn streamed_text_inside_open_code_fence(full_response: &str) -> bool {
-    let mut in_code_block = false;
-    for line in full_response.lines() {
-        if line.trim_start().starts_with("```") {
-            in_code_block = !in_code_block;
-        }
-    }
-    in_code_block
-}
-
-/// #3608: append a streamed `StreamMessage::Text` chunk to `full_response`,
-/// normalizing only the *chunk boundary* so a tool-use paragraph separator
-/// (`\n\n`, appended at the ToolUse branch) followed by a chunk that itself
-/// begins with blank lines does not accumulate into `\n\n\n\n`.
-///
-/// Narrow by construction (issue option 1): when `full_response` already ends
-/// with `\n\n` and we are NOT inside an open code fence, the chunk's leading
-/// `\n` run is trimmed before appending so the boundary collapses to a single
-/// `\n\n`. Intentional larger gaps *within* a single chunk are preserved, and
-/// blank lines inside an open code fence are never touched.
-fn append_streamed_text_chunk(full_response: &mut String, content: &str) {
-    if full_response.ends_with("\n\n") && !streamed_text_inside_open_code_fence(full_response) {
-        full_response.push_str(content.trim_start_matches('\n'));
-    } else {
-        full_response.push_str(content);
-    }
-}
-
-/// #3608: append the tool-use paragraph separator to `full_response`.
-///
-/// When a `StreamMessage::ToolUse` arrives mid-turn we trim trailing
-/// whitespace off the accumulated body and append exactly one `\n\n` so the
-/// post-tool prose starts on its own paragraph. This is the *only* boundary
-/// `append_streamed_text_chunk` keys off (a `\n\n` suffix), so the two helpers
-/// are the matched pair that composes `text → tool → text` into a single
-/// separator. Extracting the ToolUse side into its own primitive lets the
-/// regression test drive the *real* boundary instead of hand-rolling it, so a
-/// logic regression in either primitive is caught.
-///
-/// Caller keeps the surrounding `inflight_state` / `state_dirty` side effects
-/// inline — this helper is pure string composition only (no relay/watcher/
-/// ownership state, per the #3016 hot-file constraint). No-op on an empty body.
-fn append_tool_boundary_separator(full_response: &mut String) {
-    if full_response.is_empty() {
-        return;
-    }
-    let trimmed = full_response.trim_end();
-    full_response.truncate(trimmed.len());
-    full_response.push_str("\n\n");
-}
-
-#[cfg(test)]
-mod chunk_boundary_blank_line_tests {
-    use super::{
-        append_streamed_text_chunk, append_tool_boundary_separator,
-        streamed_text_inside_open_code_fence,
-    };
-
-    // #3608 regression: Text("first") → ToolUse(...) → Text("\n\nsecond") must
-    // compose `first\n\nsecond`, NOT `first\n\n\n\nsecond`.
-    //
-    // codex review on PR #3609: the boundary is now driven through the *real*
-    // production composition primitives — `append_tool_boundary_separator` for
-    // the ToolUse arm and `append_streamed_text_chunk` for the Text arm —
-    // instead of hand-rolling the trim+push_str. This ties the test to the
-    // same helpers the production loop calls, so a logic regression in either
-    // primitive (e.g. the separator no longer collapsing, or the chunk
-    // trimming dropped) fails here.
-    #[test]
-    fn tool_boundary_then_blank_leading_chunk_collapses_to_single_separator() {
-        let mut full_response = String::new();
-        append_streamed_text_chunk(&mut full_response, "first");
-        // Real ToolUse boundary primitive (production path).
-        append_tool_boundary_separator(&mut full_response);
-        // Next text chunk that itself starts with a blank line.
-        append_streamed_text_chunk(&mut full_response, "\n\nsecond");
-
-        assert_eq!(full_response, "first\n\nsecond");
-        assert!(!full_response.contains("\n\n\n"));
-    }
-
-    // #3608 call-site wiring: the full Text → ToolUse → Text composition
-    // sequence, driven end-to-end through BOTH production primitives, must
-    // yield a single paragraph separator. This is the matched-pair contract —
-    // `append_tool_boundary_separator` only ever leaves a `\n\n` suffix, which
-    // is exactly the suffix `append_streamed_text_chunk` keys off to trim the
-    // next chunk's leading blank run. If the separator primitive stops
-    // emitting `\n\n`, or the text primitive stops collapsing the boundary,
-    // the two no longer compose and this assertion fails.
-    #[test]
-    fn text_tooluse_text_sequence_composes_single_separator_via_real_primitives() {
-        let mut full_response = String::new();
-        // Text arm chunk 1.
-        append_streamed_text_chunk(&mut full_response, "first");
-        // ToolUse arm separator.
-        append_tool_boundary_separator(&mut full_response);
-        assert_eq!(full_response, "first\n\n");
-        // Text arm chunk 2 (provider re-emits its own leading blank lines).
-        append_streamed_text_chunk(&mut full_response, "\n\nsecond");
-
-        assert_eq!(full_response, "first\n\nsecond");
-        assert!(!full_response.contains("\n\n\n"));
-    }
-
-    // `append_tool_boundary_separator` is a no-op on an empty body (mirrors the
-    // production `if !full_response.is_empty()` guard around the ToolUse arm):
-    // a tool call before any assistant text must not seed a leading separator.
-    #[test]
-    fn tool_boundary_separator_is_noop_on_empty_body() {
-        let mut full_response = String::new();
-        append_tool_boundary_separator(&mut full_response);
-        assert_eq!(full_response, "");
-    }
-
-    // `append_tool_boundary_separator` collapses pre-existing trailing
-    // whitespace to exactly one `\n\n` separator (idempotent boundary): a body
-    // that already ends in blank lines must not stack a second separator.
-    #[test]
-    fn tool_boundary_separator_collapses_trailing_whitespace() {
-        let mut full_response = String::from("first\n\n");
-        append_tool_boundary_separator(&mut full_response);
-        assert_eq!(full_response, "first\n\n");
-
-        let mut trailing_spaces = String::from("first  \n  \n");
-        append_tool_boundary_separator(&mut trailing_spaces);
-        assert_eq!(trailing_spaces, "first\n\n");
-    }
-
-    // A single intentional `\n\n` separator with a non-blank-leading follow-up
-    // chunk must pass through untouched (no over-trimming).
-    #[test]
-    fn tool_boundary_then_plain_chunk_is_unchanged() {
-        let mut full_response = String::from("first\n\n");
-        append_streamed_text_chunk(&mut full_response, "second");
-        assert_eq!(full_response, "first\n\nsecond");
-    }
-
-    // Without a preceding `\n\n` boundary, a chunk's own leading newlines are
-    // preserved verbatim — the normalization is boundary-only.
-    #[test]
-    fn no_boundary_separator_preserves_chunk_leading_newlines() {
-        let mut full_response = String::from("first");
-        append_streamed_text_chunk(&mut full_response, "\nsecond");
-        assert_eq!(full_response, "first\nsecond");
-
-        let mut single_nl = String::from("first\n");
-        append_streamed_text_chunk(&mut single_nl, "\nsecond");
-        // ends with only one `\n`, so not a `\n\n` boundary → no trim.
-        assert_eq!(single_nl, "first\n\nsecond");
-    }
-
-    // Intentional larger gaps INSIDE a single chunk are preserved; only the
-    // cross-chunk boundary is normalized.
-    #[test]
-    fn intentional_gap_within_single_chunk_is_preserved() {
-        let mut full_response = String::new();
-        append_streamed_text_chunk(&mut full_response, "a\n\n\n\nb");
-        assert_eq!(full_response, "a\n\n\n\nb");
-    }
-
-    // Blank lines inside an OPEN code fence must not be touched: when the
-    // accumulated body ends inside a fence, the incoming chunk's leading
-    // newlines are intentional fenced content and are preserved.
-    #[test]
-    fn open_code_fence_preserves_chunk_leading_newlines() {
-        let mut full_response = String::from("```\ncode line\n\n");
-        assert!(streamed_text_inside_open_code_fence(&full_response));
-        append_streamed_text_chunk(&mut full_response, "\n\nmore code");
-        // Inside the fence: leading blank lines kept verbatim.
-        assert_eq!(full_response, "```\ncode line\n\n\n\nmore code");
-    }
-
-    // A closed fence followed by prose normalizes normally (we are no longer
-    // inside the fence at the boundary).
-    #[test]
-    fn closed_code_fence_then_boundary_normalizes() {
-        let mut full_response = String::from("```\ncode\n```\n\n");
-        assert!(!streamed_text_inside_open_code_fence(&full_response));
-        append_streamed_text_chunk(&mut full_response, "\n\nafter");
-        assert_eq!(full_response, "```\ncode\n```\n\nafter");
-    }
 }
 
 fn turn_bridge_stream_wait_duration(
@@ -2057,7 +1871,7 @@ pub(super) fn spawn_turn_bridge(
                             // #3608: normalize the chunk boundary so a tool-use
                             // `\n\n` separator + a chunk that itself begins with
                             // blank lines does not accumulate into `\n\n\n\n`.
-                            append_streamed_text_chunk(&mut full_response, &content);
+                            chunk_compose::append_streamed_text_chunk(&mut full_response, &content);
                             if (watcher_owns_assistant_relay
                                 && watcher_relay_available_for_turn)
                                 || standby_relay_owns_output
@@ -2390,7 +2204,7 @@ pub(super) fn spawn_turn_bridge(
                                 // `append_streamed_text_chunk`). inflight_state
                                 // / state_dirty stay inline (hot-file #3016:
                                 // only full_response composition is extracted).
-                                append_tool_boundary_separator(&mut full_response);
+                                chunk_compose::append_tool_boundary_separator(&mut full_response);
                                 inflight_state.full_response = full_response.clone();
                                 state_dirty = true;
                             }
